separated documantion in blog, docs and examples; new guide and blog article
This update separates the documentation of this library into blog, docs and examples. The blog contains articles of the author about experiences, ideas and take aways regarding this library. The docs folder contatins the documenation about best practices, the features of this library. The examples folder contains examples for use cases.